Make separate deep-equals.js module
