+++ .prettierignore
@@ -303,8 +303,10 @@ packages/labs/ssr-react/enable-lit-ssr.*
 packages/labs/task/development/
 packages/labs/task/test/
 packages/labs/task/node_modules/
+packages/labs/task/deep-equals.*
 packages/labs/task/index.*
 packages/labs/task/task.*
+
 packages/labs/testing/index.*
 packages/labs/testing/fixtures.*
 packages/labs/testing/web-test-runner-ssr-plugin.*
