+++ packages/labs/task/.gitignore
@@ -1,5 +1,6 @@
 /development/
 /test/
 /node_modules/
+/deep-equals.*
 /index.*
 /task.*
