+++ packages/labs/task/package.json
@@ -22,6 +22,11 @@
       "development": "./development/index.js",
       "default": "./index.js"
     },
+    "./deep-equals.js": {
+      "types": "./development/deep-equals.d.ts",
+      "development": "./development/deep-equals.js",
+      "default": "./deep-equals.js"
+    },
     "./task.js": {
       "types": "./development/task.d.ts",
       "development": "./development/task.js",
@@ -31,6 +36,7 @@
   "files": [
     "/development/",
     "!/development/test/",
+    "/deep-equals.{d.ts,d.ts.map,js,js.map}",
     "/index.{d.ts,d.ts.map,js,js.map}",
     "/task.{d.ts,d.ts.map,js,js.map}"
   ],
@@ -92,6 +98,7 @@
         "../../../rollup-common.js"
       ],
       "output": [
+        "deep-equals.js{,.map}",
         "index.js{,.map}",
         "task.js{,.map}",
         "test/**/*.js{,.map}"
